rewards and crafts doesn't always return full items

+++ schema.js
@@ -28,7 +28,7 @@ module.exports = `
 
   type TaskItem {
       item: Item!
-      count: Int!
+      count: Float!
   }
 
   type Barter {
